requirements.txt
Django==3.0.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-redis==4.12.1
channels==2.3.0
channels-redis==3.0.0
celery==4.4.6
uvicorn==0.11.1
six==1.15.0
